feat(KSegmentedControl): updates according to PR [KHCP-6514]
