+++ src/utilities/index.ts
@@ -1,3 +1,2 @@
 export { copyTextToClipboard } from './copyTextToClipboard'
 export { debounce } from './debounce'
-export * from './segmentControlValidators'
